+++ web/bin/takode.ts
@@ -17,7 +17,6 @@ import {
   stripGlobalFlags,
 } from "./takode-core.js";
 import { handleFileResolve } from "./takode-file-resolve.js";
-import { handleContextDoctor } from "./takode-context-diagnostics.js";
 import { printCommandHelp, printUsage } from "./takode-help.js";
 import { handleLease } from "./takode-lease.js";
 import { handleExport, handleGrep, handleLogs, handlePeek, handleRead, handleScan } from "./takode-message-commands.js";
@@ -69,7 +68,6 @@ try {
     ["search", {}],
     ["info", {}],
     ["leader-context-resume", {}],
-    ["context-doctor", {}],
     ["file-resolve", {}],
     ["tasks", {}],
     ["timers", {}],
@@ -147,9 +145,6 @@ try {
       break;
     case "leader-context-resume":
       await handleLeaderContextResume(base, args);
-      break;
-    case "context-doctor":
-      await handleContextDoctor(base, args);
       break;
     case "file-resolve":
       await handleFileResolve(base, args);